wc: moved components to deprecated #TASK-2051
