+++ src/webcomponents/study/deprecated/variable-update.js
